SPT-4113 possible fix
Trying my hand at updating this.  The update to history fixes the issue
we were having, and I added an assert to the mock test but I feel like
Im missing something else to actually make that test pass.  Cant wrap my
head around how it will grab an application history revision.

I welcome any pointers on what I could have done better.  I wanted to
find a way to keep a copy of the current app_revision to pass down the
list.  Then we could check the app version number instead of having to
pull the same version of the app over and over.  Couldnt see a way to do
that without a much larger overhaul.  Hopefully I'm wrong and there is a
more efficient way to do it.