Update docs to use the latest version of Sphinx
We are using a really old version and the build is failing locally
when I attempt to build the docs. Bumping to the latest fixes this and
doesn't appear to break anything.

We also need to build the docs using Python3 since PyDriver is now built
against Python 3.0, attempting to build using Python2 will result in a
build error when generating the docs.