Updating Python Version in setup.py
This got bumped to the wrong version while the release team was testing
out a GitHub Action for cutting a new PyDriver release as part of the
platform build process. The version number in the documentation should
always match what we have released as the latest in Pypi and GitHub.