SPT-16723 - Fix Python Docs (#178)
* Update docs to use the latest version of Sphinx

We are using a really old version and the build is failing locally
when I attempt to build the docs. Bumping to the latest fixes this and
doesn't appear to break anything.

We also need to build the docs using Python3 since PyDriver is now built
against Python 3.0, attempting to build using Python2 will result in a
build error when generating the docs.

* Adding config file for ReadTheDocs

This will control what versions of Sphinx and Python are used to build
the documentation so that we don't run into issues with repeatable
builds of the documentation.

* Updating Python Version in setup.py

This got bumped to the wrong version while the release team was testing
out a GitHub Action for cutting a new PyDriver release as part of the
platform build process. The version number in the documentation should
always match what we have released as the latest in Pypi and GitHub.